requirements.txt
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
boto3==1.26.98
botocore==1.29.98
certifi==2022.12.7
charset-normalizer==3.1.0
comm==0.1.3
debugpy==1.6.6
decorator==5.1.1
executing==1.2.0
ffmpeg-python==0.2.0
filelock==3.10.4
future==0.18.3
idna==3.4
importlib-metadata==6.1.0
ipykernel==6.22.0
ipython==8.11.0
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
jupyter_client==8.1.0
jupyter_core==5.3.0
llvmlite==0.39.1
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
more-itertools==9.1.0
mpmath==1.3.0
nest-asyncio==1.5.6
networkx==3.0
numba==0.56.4
numpy==1.23.5
openai-whisper==20230314
packaging==23.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.1.1
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pyzmq==25.0.2
regex==2023.3.23
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
stack-data==0.6.2
sympy==1.11.1
tiktoken==0.3.1
torch==2.0.0
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
zipp==3.15.0
